+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/rsc
 
+## 3.0.37
+
+### Patch Changes
+
+- ai@7.0.37
+
 ## 3.0.36
 
 ### Patch Changes